fsperf: fixup clean results
The script wasn't quite right for adding arguments, and add a few
arguments so we can easily clean up botched runs.

+++ src/clean-results.py
@@ -2,11 +2,18 @@ from ResultData import *
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 import argparse
+import sys
 
 parser = argparse.ArgumentParser()
-parser.add_argument("labels", nargs='+', type=str,
+parser.add_argument("--labels", nargs='*', type=str, default=[],
                     help="Labels to delete the results for")
+parser.add_argument("--config", type=str,
+                    help="Configs to delete the results for")
 args = parser.parse_args()
+
+if not args.labels and not args.config:
+    print("Must specify either labels or configs to delete from")
+    sys.exit(1)
 
 engine = create_engine('sqlite:///fsperf-results.db')
 Session = sessionmaker()
@@ -19,6 +26,16 @@ for p in args.labels:
             outerjoin(DbenchResult).\
             outerjoin(TimeResult).\
             filter(Run.purpose == p).all()
+    for r in results:
+        session.delete(r)
+    session.commit()
+
+if args.config is not None:
+    results = session.query(Run).\
+            outerjoin(FioResult).\
+            outerjoin(DbenchResult).\
+            outerjoin(TimeResult).\
+            filter(Run.config == args.config).all()
     for r in results:
         session.delete(r)
     session.commit()
